// dict/api/user/types.ts
/* eslint-disable */
// tslint:disable
/**
 * MICRO SERVICES CURD OPERATIONS
 * 
 *
 * OpenAPI spec version: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface UserApi {
	createTopic: (request: Api.BODYDATA | undefined) => Promise<CreateTopicResponse>
	deleteTopic: (topicId: number) => Promise<DeleteTopicResponse>
	getTopic: () => Promise<GetTopicResponse>
	updateTopic: (topicId: number, request: Api.BODYDATA | undefined) => Promise<UpdateTopicResponse>
}

export type CreateTopicResponse = CreateTopic201Response | CreateTopic400Response

export interface CreateTopic201Response {
	status: 201
	body: Api.ApiResponse
	headers?: never
}

export interface CreateTopic400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export type DeleteTopicResponse = DeleteTopic201Response | DeleteTopic400Response

export interface DeleteTopic201Response {
	status: 201
	body: Api.ApiResponse
	headers?: never
}

export interface DeleteTopic400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export type GetTopicResponse = GetTopic201Response | GetTopic400Response

export interface GetTopic201Response {
	status: 201
	body: Api.BODYDATA[]
	headers?: never
}

export interface GetTopic400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export type UpdateTopicResponse = UpdateTopic201Response | UpdateTopic400Response

export interface UpdateTopic201Response {
	status: 201
	body: Api.ApiResponse
	headers?: never
}

export interface UpdateTopic400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}